platform/requirements.txt
Flask==2.3.2
Flask-Mail==0.9.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
python-dateutil==2.8.2
pyotp==2.9.0
Werkzeug==2.3.7
pymongo==4.3.3
